Version 1.0 | Scaffolding
Barreling ahead with lots of code and development just to get this program off the ground.

I've arrived at a point know where a lot of architectural changes are being made, and I'm being to worry if not committing former decisions would be a good idea.

Thus, we have our first commit. About to redesign the program - hopefully this goes well.