Version 1.3 | Scoring System Online
Program now accepts the correct answers is not sensitive to white space (before or after the string, you still have to match the correct answers, character for character). It is not case insensitive at this time.

Program also tracks your accuracy / score. This is great for qualifying your success with the quiz. This can be utilized for data and analytics in future editions.