Version 1.4 | Answer Case Insensitivity
Program is now case insensitive when it comes to the answer comparison mechanism between the user input and the predefined answers in the ,json file. User laziness has been greatly enabled.

Only thing left is allowing the user to be lazy about answer formatting.

For example, allowing "cloudstorage" to be an accepted answer, when the answer is technically "cloud storage" or "Cloud Storage" instead.