Version 1.6 | Program Loop
Implemented a program loop. Now the program will not simply end when you're done with your question set. You have the option of re-running the same question set or running a different one (or terminating the program).

Additionally, the procedural code was getting to be too cumbersome. Modularized the code some, which reduces the number of lines of code in the program.