Version 1.9 | Order Agnostic
The program is now answer order agnostic. Well... sort of.

Each question in their JSON file has been given a new property, "OrderAgnostic." For questions where this property is set to true, it does not matter in which order you decide to write the strings of the correct answers in. So long as all of the correct answers are there, all the strings are present in your answer - your input will be correct.

However, if "OrderAgnostic" is false - then yes, you may be marked wrong if your answers are not in the correct order. This is even if you have all the correct strings there.

For example, one of the questions is,

"Of the six database options on Google Cloud Platform, which option is SQL-based and which is relational (in that respective order)?"

This question stipulates that the answer for the option which is based in SQL be written before the option that is relational. The order matters here, as this is how the question is written (and writing it any other way would be a pain).

Think the last thing I'll crank out today is getting the correct answers printed at the end of the review session.

KBAI. : D