Version 4.0 | Windows Startup
I have successfully implemented a solution to allow the .py script to start up with Windows.

This solution was developed on Windows 11 and in my environment, so this functionality may not be accessible for a multitude of OS, configurations or unique environments.

When I turn my focus to making this software more portable, I'll address that.