Version 4.2 | Randomization
Implemented two review session question presentation modes: (1) Random and (2) Sequence.

Sequence is the standard, where the questions will be presented in the same order every time - the way in which the JSON file was written.

Random will shuffle that order and present the questions to you in a randomized way each time.