Version 4.3 | Option Randomization
For the same reason that questions would become too predictable if presented in the same order every time, I felt that the question options would suffer in the same way.

So I've redesigned the JSON structure for the questions file. They now feature a Options array, of which the options can be randomized once passed into the Python script. I also refactored the Python script to accommodate this new logic.

At the moment, there's an existing issue I am fixing. Some of the answers to these questions are getting complex - featuring multiple sentence fragments in the answer. Well.. those fragments in English need to be connected using commas, which just so happens to be a delimiter I use in the program... So yeah.

Fixing that atm. You can see this process occurring in the test.json file (Mistakes Breakdown).

Also, I finally confirmed that the rerun function doesn't work yet. I'll have to fix that too.