Version 5.1 | KB Batch 2
Added a new knowledge base to the program:

1. Cloud VPN

Additionally, I just finished an entire round of QA testing for all existing knowledge bases (no testing yet for the cloud_vpn.json file though). Thank goodness, there was a lot wrong with them. I essentially uncovered a persistent bug in the program.

The program is able to successfully parse questions where there is only one correct answer and it has a comma. In the event that there is more than one correct answer, and one of them has a comma - the script fails and the question is marked incorrect. I think the entire comparison logic for the review_session() function is due for a refactor - I just don't have the time and energy to take care of that now.

The easier solution for the time being is to re-design answers with as few commas as possible. I've since done so to avoid issues - however, this should probably be tackled at some point in the future.

A lot of the questions were also not marked as order agnostic when they were supposed to be. Contemplating a report on the order agnosticism status during the mistakes breakdown to aid in the debugging process in future.

Finally, got the "Practice New Section" feature working.