Version 5.2 | Cloud Service Models KB Edits
Made minor edits to the cloud_service_models.json. I need to add a default of "OrderAgnostic" to true.

However, I'm committing now because I have some big changes I want to make to the program. I'm contemplating switching the whole thing to MC and simply allowing the user to type the answer for fun if they'd like (for rote memorization, you can simply press enter or anything to bypass it).

Here we go!