Version 6.0 | Multiple Choice Program
The program has now officially been reconfigured to a multiple choice answer system. This will remove a lot of the prior frustrations with having to type out long strings and make zero mistakes in order to get questions that you already know correct. It was irritating knowing the answer to a question and getting marked wrong because I swapped two consonants around.

This refactoring may have also shrunk the size of the program and added more modularization, which is kew.

I've still included an input statement, allowing the user to type out the string to assist with rote learning. You'll notice that said data isn't even stored in a variable. It serves no purpose other than to support the user. It's but a mere fantasy! BEGONE! :smiling_imp: