Version 6.1 | Random & Mistakes Repair
Partially repaired the randomization functionality. The random feature does randomize the order of the questions, just not the order of the answers.

The latter will be a complicated problem to solve, as the new architecture is predicated on fixed answer IDs in order to work. I'd have to innovate further to resolve this issue. We'll leave it be for you.

Additionally, the mistakes_breakdown feature has been repaired - now printing properly at the end of the session and being written to the data_log.