Version 7.1 | First Replacement
The json_redesign script did well for the most part. There were a couple "CorrectAnswerIDs" that were missing and Question 20 was missing the "T" option.

I'll have to dig into why this happened some other time. I'm happy enough with using diffchecker to assess the quality, making the minor adjustments I need - and replacing the original knowledge base files in the main directory.

I value rapid development over polish at this time. Looking to run this script on all the knowledge base .json files to finish updating to the new MC model. It's imperative I am not bogged down by these technical components, as my biggest priority is simply studying for the GCP ACE exam.

So now, two knowledge bases / sections are available for review in the new version of the program.

1. "Storage and Database Options"
2. "Google Cloud Global Infrastructure"

I recognize that the quality of the review sessions are largely impacted by this large update. Many of the questions frankly suck for the purpose of challenging one's self. Questions that were fill in the blank have been trivialized by having only one option to select. Other recall questions suffer from the same effect. At some point, I'll have to either re-design those questions or refactor the program to accommodate those kinds of questions again. Not sure which option I'll choose.

Oh also, happy to know as a result of my work on json_redesign_script.py - that you do not need to run a function called main for your script to work. The semantics of such a function is not important to the Python program like it is in C++.