Version 7.2 | Second Replacement
Successfully converted another knowledge base over to the new schema. I had to make changes to the redesign script, as it wasn't working as is... Which begs the question - how did it work in the first place in the First Replacement commit? Not sure. I imagine it'll work just fine from here. We'll see.

Also - I didn't realize the Python script was inputting all CorrectAnswerIDs as lists, whether or not they are MultipleCorrectAnswers. For this reason, I assumed I had to go back to some of the other .json files and edit them. Convert the single standing integer values into a a list with 1 integer value inside of it.

Turns out I don't. My code is able to recognize both structures, whether the integer is wrapped in a list or not. Decided to leave files like google_cloud_global_infrastructure.json be then as a testament to my greatness. :sunglasses: