7.3 | Lost in the Sauce
Okay, maybe the previous JSON conversion featured a JSON file that somehow deviated from the original JSON structure I developed the others with? We're back to the same script logic. Wouldn't work with the new version of json_resdesign_script.py that I had created in the last commit.

Hopefully this is the last edition. All the other JSON files look like the networking_services.json file. Anyhoo, let's see if we can punch out the rest of these conversions and be done with it.